# contracts/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "KillProcesslistRequest": {
        "properties": {
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "title": "KillProcesslistRequest",
        "type": "object"
      },
      "KillProcesslistResponse": {
        "properties": {
          "checked_at": {
            "title": "Checked At",
            "type": "string"
          },
          "killed": {
            "title": "Killed",
            "type": "boolean"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "required": [
          "checked_at",
          "killed",
          "notes"
        ],
        "title": "KillProcesslistResponse",
        "type": "object"
      },
...
      "ProcesslistEntryResponse": {
        "properties": {
          "command": {
            "title": "Command",
            "type": "string"
          },
          "db": {
            "title": "Db",
            "type": "string"
          },
          "host": {
            "title": "Host",
            "type": "string"
          },
          "info": {
            "title": "Info",
            "type": "string"
          },
          "process_id": {
            "title": "Process Id",
            "type": "integer"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "time_seconds": {
            "title": "Time Seconds",
            "type": "integer"
          },
          "trx_started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trx Started At"
          },
          "user": {
            "title": "User",
            "type": "string"
          }
        },
        "required": [
          "process_id",
          "user",
          "host",
          "db",
          "command",
          "time_seconds",
          "state",
          "info",
          "trx_started_at"
        ],
        "title": "ProcesslistEntryResponse",
        "type": "object"
      },
      "ProcesslistSnapshotResponse": {
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/ProcesslistEntryResponse"
            },
            "title": "Entries",
            "type": "array"
          },
          "snapshot_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot At"
          }
        },
        "required": [
          "snapshot_at",
          "entries"
        ],
        "title": "ProcesslistSnapshotResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "db-monitor-api",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/instances/{instance_id}/processlist": {
      "get": {
        "operationId": "get_instance_processlist_instances__instance_id__processlist_get",
        "parameters": [
          {
            "in": "path",
            "name": "instance_id",
            "required": true,
            "schema": {
              "title": "Instance Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "user",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User"
            }
          },
          {
            "in": "query",
            "name": "host",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Host"
            }
          },
          {
            "in": "query",
            "name": "command",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Command"
            }
          },
          {
            "in": "query",
            "name": "min_time_seconds",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Time Seconds"
            }
          },
          {
            "in": "query",
            "name": "collected_after",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Collected After"
            }
          },
          {
            "in": "query",
            "name": "collected_before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Collected Before"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcesslistSnapshotResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Instance Processlist"
      }
    },
    "/instances/{instance_id}/processlist/{process_id}/kill": {
      "post": {
        "operationId": "kill_instance_process_instances__instance_id__processlist__process_id__kill_post",
        "parameters": [
          {
            "in": "path",
            "name": "instance_id",
            "required": true,
            "schema": {
              "title": "Instance Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "process_id",
            "required": true,
            "schema": {
              "title": "Process Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/KillProcesslistRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KillProcesslistResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Kill Instance Process"
      }
    }
  }
}